+++ tmp/tmp.go
@@ -31,6 +31,12 @@ type Temp struct {
 // Dir is default temporary directory
 var Dir = "/tmp"
 
+// DefaultDirPerms is default permissions for directories
+var DefaultDirPerms os.FileMode = 0750
+
+// DefaultFilePerms is default permissions for files
+var DefaultFilePerms os.FileMode = 0640
+
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 // NewTemp create new Temp structure
@@ -67,7 +73,7 @@ func (t *Temp) MkDir(args ...string) (string, error) {
 	}
 
 	tmpDir := getTempName(t.Dir, name)
-	err := os.MkdirAll(tmpDir, 0750)
+	err := os.MkdirAll(tmpDir, DefaultDirPerms)
 
 	if err != nil {
 		return "", err
@@ -87,7 +93,7 @@ func (t *Temp) MkFile(args ...string) (*os.File, string, error) {
 	}
 
 	tmpFile := getTempName(t.Dir, name)
-	fd, err := os.OpenFile(tmpFile, os.O_RDWR|os.O_CREATE, 0640)
+	fd, err := os.OpenFile(tmpFile, os.O_RDWR|os.O_CREATE, DefaultFilePerms)
 
 	if err != nil {
 		return nil, "", err