+++ changelog.md
@@ -6,6 +6,7 @@
 * `[terminal]` Added hints support from new version of `go.linenoise`
 * `[fmtc]` Light colors tag (`-`) support
 * `[usage]` Using dark grey color for option values and example description
+* `[tmp]` Added `DefaultDirPerms` and `DefaultFilePerms` global variables for permissions customization
 
 #### v3.4.2